Moved bosonic bases and watson_hamiltonian.py
