+++ qiskit_nature/__init__.py
@@ -159,7 +159,7 @@ from .bosonic_operator import BosonicOperator
 from .exceptions import QiskitNatureError
 from .fermionic_operator import FermionicOperator
 from .mp2info import MP2Info
-from .watson_hamiltonian import WatsonHamiltonian
+from qiskit_nature.drivers.watson_hamiltonian import WatsonHamiltonian
 
 __all__ = ['QiskitNatureError',
            'WatsonHamiltonian',
